Add new hyperbezier math
This commit adds a first implementation of the new hyperbezier curve. It doesn't yet offer much help in determining the raw parameters, but does produce nice Bézier paths through curve fitting.
