+++ Cargo.toml
@@ -3,7 +3,7 @@ name = "spline"
 version = "0.3.0"
 license = "MIT/Apache-2.0"
 authors = ["Raph Levien <raph.levien@gmail.com>"]
-edition = "2018"
+edition = "2021"
 keywords = ["graphics", "curve", "curves", "bezier", "spline"]
 repository = "https://github.com/linebender/spline"
 description = "A spline for interactive 2D curve design"
@@ -11,7 +11,7 @@ readme = "README.md"
 categories = ["graphics"]
 
 [dependencies]
-kurbo = "0.9"
+kurbo = "0.11"
 serde_ = { version = "1.0.117", package="serde", features = ["derive"], optional = true }
 
 [dev-dependencies]
